Improved list code

+++ src/gfx/renderer.c
@@ -120,7 +120,7 @@ void render_face(const struct KCR_Display* display, const struct KCR_Scene* scen
 }
 
 void kcr_render(const struct KCR_Display *display, const struct KCR_Scene *scene) {
-    for (int idx = 0; idx < kcr_list_length(scene->mesh->faceList); idx++) {
+    for (size_t idx = 0; idx < kcr_list_length(scene->mesh->faceList); idx++) {
         const struct KCR_Face* face = &scene->mesh->faceList[idx];
         const struct TransformedFace transformedFace = transformFace(face, scene->mesh, &scene->mesh->rotation);
